+++ apps/web/src/components.d.ts
@@ -55,16 +55,6 @@ declare module 'vue' {
     ProjectPrograms: typeof import('./components/project/details/ProjectPrograms.vue')['default']
     ProjectProgramsList: typeof import('./components/project/details/ProjectProgramsList.vue')['default']
     ProjectSideNav: typeof import('./components/project/details/ProjectSideNav.vue')['default']
-    RegisterDetailActivity: typeof import('./components/identification/details/RegisterDetailActivity.vue')['default']
-    RegisterDetailElement: typeof import('./components/identification/details/RegisterDetailElement.vue')['default']
-    RegisterDetailLocalisation: typeof import('./components/identification/details/RegisterDetailLocalisation.vue')['default']
-    RegisterDetailManual: typeof import('./components/identification/details/RegisterDetailManual.vue')['default']
-    RegisterDetailSelectSize: typeof import('./components/identification/details/RegisterDetailSelectSize.vue')['default']
-    RegisterDetailsInfos: typeof import('./components/identification/details/RegisterDetailsInfos.vue')['default']
-    RegisterDetailSiret: typeof import('./components/identification/details/RegisterDetailSiret.vue')['default']
-    RegisterDetailSize: typeof import('./components/identification/details/RegisterDetailSize.vue')['default']
-    RegisterDetailTag: typeof import('./components/identification/details/RegisterDetailTag.vue')['default']
-    RegisterSiret: typeof import('./components/identification/RegisterSiret.vue')['default']
     ResultHeader: typeof import('./components/questionnaire/result/list/ResultHeader.vue')['default']
     ResultList: typeof import('./components/questionnaire/result/list/ResultList.vue')['default']
     ResultListInTabs: typeof import('./components/questionnaire/result/list/ResultListInTabs.vue')['default']
@@ -107,11 +97,18 @@ declare module 'vue' {
     TeeListNoResults: typeof import('./components/element/TeeListNoResults.vue')['default']
     TeeMatomo: typeof import('./components/TeeMatomo.vue')['default']
     TeeNoResult: typeof import('./components/element/TeeNoResult.vue')['default']
+    TeeProfileActivity: typeof import('./components/identification/details/TeeProfileActivity.vue')['default']
+    TeeProfileDetailElement: typeof import('./components/identification/details/TeeProfileDetailElement.vue')['default']
+    TeeProfileDetails: typeof import('./components/identification/details/TeeProfileDetails.vue')['default']
+    TeeProfileLocalisation: typeof import('./components/identification/details/TeeProfileLocalisation.vue')['default']
+    TeeProfileSiret: typeof import('./components/identification/details/TeeProfileSiret.vue')['default']
+    TeeProfileSize: typeof import('./components/identification/details/TeeProfileSize.vue')['default']
     TeeProjectButton: typeof import('./components/project/TeeProjectButton.vue')['default']
     TeeQuestionnaire: typeof import('./components/questionnaire/TeeQuestionnaire.vue')['default']
     TeeQuestionnaireResult: typeof import('./components/questionnaire/TeeQuestionnaireResult.vue')['default']
     TeeRegisterCTA: typeof import('./components/identification/TeeRegisterCTA.vue')['default']
     TeeRegisterModal: typeof import('./components/identification/TeeRegisterModal.vue')['default']
+    TeeRegisterSiret: typeof import('./components/identification/TeeRegisterSiret.vue')['default']
     TeeRegisterSiretBar: typeof import('./components/identification/TeeRegisterSiretBar.vue')['default']
     TeeSpinner: typeof import('./components/element/TeeSpinner.vue')['default']
     TeeTabs: typeof import('./components/element/TeeTabs.vue')['default']
